gRPC - fixes and minor usability improvements
